klaudbiusz/pyproject.toml
[project]
name = "klaudbiusz"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "fire>=0.7.1",
    "coloredlogs>=15.0.1",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "joblib>=1.4.0",
    "anthropic>=0.40.0",
    "mlflow>=2.15.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pyright>=1.1.406",
]
